Formatting and fixes

+++ back-end/src/optimizer.rs
@@ -45,20 +45,13 @@ fn load_bundled_dir(bundled_dir: &Dir, target: PathBuf) -> io::Result<()> {
 }
 
 pub fn optimize_settings(config: &OptimizerConfig, user_profile: &UserProfile) -> io::Result<()> {
-    let ssbu_settings_path = config
-        .local_data
-        .emu_folder
-        .as_ref()
-        .ok_or(io_error!(NotFound, "Emulator data folder not found"))?
-        .join("config")
+    let config_folder = config.emu_filesystem.config_folder.as_ref();
+    let ssbu_settings_path = config_folder
+        .ok_or(io_error!(NotFound, "Emulator config folder not found"))?
         .join("custom");
     load_bundled_dir(&BUNDLED_SSBU_SETTINGS, ssbu_settings_path)?;
-    let main_config_settings_path = config
-        .local_data
-        .emu_folder
-        .as_ref()
-        .ok_or(io_error!(NotFound, "Emulator data folder not found"))?
-        .join("config")
+    let main_config_settings_path = config_folder
+        .ok_or(io_error!(NotFound, "Emulator config folder not found"))?
         .join("qt-config.ini");
     let mut main_config = Ini::load_from_file_noescape(main_config_settings_path.as_path())
         .ok()
@@ -96,7 +89,8 @@ pub fn optimize_mods(
     user_profile: &UserProfile,
     advanced_options: Vec<AdvancedOption>,
 ) -> io::Result<()> {
-    let sdmc_path = config.emu_filesystem
+    let sdmc_path = config
+        .emu_filesystem
         .sdmc_folder
         .as_ref()
         .ok_or(io_error!(NotFound, "Unable to find sdmc directory"))?;
@@ -105,23 +99,18 @@ pub fn optimize_mods(
         .join("contents")
         .join(SSBU_TITLE_ID);
 
-    if advanced_options.contains(&AdvancedOption::CleanSkyline) {
-        if skyline_path.is_dir() {
-            log::info!("Removing skyline files...");
-            fs::remove_dir_all(skyline_path.as_path())?;
-        }
+    if advanced_options.contains(&AdvancedOption::CleanSkyline) && skyline_path.is_dir() {
+        log::info!("Removing skyline files...");
+        fs::remove_dir_all(skyline_path.as_path())?;
     }
 
-    let arc_config_path = config.get_arc_config_folder(&user_profile)?;
-    let arc_mods_path = sdmc_path
-        .join("ultimate");
+    let arc_config_path = config.get_arc_config_folder(user_profile)?;
+    let arc_mods_path = sdmc_path.join("ultimate");
 
-    if advanced_options.contains(&AdvancedOption::CleanArc) {
-        if arc_mods_path.is_dir() {
-            log::info!("Removing arcropolis files...");
-            fs::remove_dir_all(arc_config_path.as_path())?;
-            fs::remove_dir_all(arc_mods_path.as_path())?;
-        }
+    if advanced_options.contains(&AdvancedOption::CleanArc) && arc_mods_path.is_dir() {
+        log::info!("Removing arcropolis files...");
+        fs::remove_dir_all(arc_config_path.as_path())?;
+        fs::remove_dir_all(arc_mods_path.as_path())?;
     }
 
     let arc_mods_path = arc_mods_path.join("mods");
@@ -136,7 +125,7 @@ pub fn optimize_mods(
 }
 
 pub fn optimize_save(config: &OptimizerConfig, user_profile: &UserProfile) -> io::Result<()> {
-    let save_file_path = config.get_save_folder(&user_profile)?;
+    let save_file_path = config.get_save_folder(user_profile)?;
     load_bundled_dir(&BUNDLED_SAVE_DATA, save_file_path)?;
     Ok(())
 }
